dwarf: pull abbrev_* consts into a sized enum AbbrevKind
Add initial support for emitting debug info for local variables
(on the stack) within a subprogram. However, to do this, I had to
move writing to debug info back to `CodeGen` from `Emit` as we
require the operand resolved at the callsite (plus its type). This
is crucial to be done at callsite as, if postponed, we might end up
with a `dead` mcv.